requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pipecat-ai[all,livekit]
ollama
faster-whisper
kokoro
numpy
livekit-api
livekit
livekit-agents
livekit-plugins-openai
livekit-plugins-google
livekit-plugins-deepgram
livekit-plugins-silero
pyjwt
tenacity
